Обновление вызовов методов ядра, неймспейсы
